update_10061403

+++ src/ScaleTruckController.cpp
@@ -83,7 +83,7 @@ void* ScaleTruckController::lanedetectInThread() {
   centerLine_ = laneDetector_.display_img(camImageTmp, waitKeyDelay_, viewImage_);
   float weight = (centerLine_ - centerErr_)/centerErr_*(-1.0f);
 
-  AngleDegree_ = atan(width2dist_*weight/dist_)*180.0f/M_PI;
+  AngleDegree_ = atanf(width2dist_*weight/dist_)*(180.0f/M_PI);
 
   //AngleDegree_ = weight * AngleMax_; // -1 ~ 1 
 }
